esp32_simulator/requirements.txt
firebase-admin
requests
asyncio
python-dotenv==1.0.1
